+++ test/Unit/Ads/CustomFeaturesTest.ts
@@ -5,6 +5,9 @@ import { toAbGroup } from 'Core/Models/ABGroup';
 import { Platform } from 'Core/Constants/Platform';
 import * as sinon from 'sinon';
 
+// Only for the slider experiment
+const invalidABGroup: number = 5;
+
 describe('CustomFeatures', () => {
 
     describe('isExampleGameId', () => {
@@ -125,15 +128,12 @@ describe('CustomFeatures', () => {
         });
     });
 
-    describe('isPartOfPhaseTwoRollout', () => {
+    describe('isZyngaWordsWithFriends', () => {
         const tests: {
             gameId: string;
             expected: boolean;
         }[] = [{
             gameId: '2895988', // WWF
-            expected: true
-        }, {
-            gameId: '2988443', // Zynga Solitaire
             expected: true
         }, {
             gameId: 'scott',
@@ -142,7 +142,7 @@ describe('CustomFeatures', () => {
 
         tests.forEach(t => {
             it('should match the expected value', () => {
-                const value = CustomFeatures.isPartOfPhaseTwoLoadRollout(t.gameId);
+                const value = CustomFeatures.isZyngaWordsWithFriends(t.gameId);
                 assert.equal(value, t.expected);
             });
         });
@@ -286,60 +286,18 @@ describe('CustomFeatures', () => {
     });
 
     describe('isSliderEndScreenEnabled', () => {
-        it('should return false for iOS 9.0 if ab group is 5 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '9.0', Platform.IOS);
+        it(`should return false for iOS 9.0 if ab group is ${invalidABGroup} and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.`, () => {
+            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(invalidABGroup), '547145938', '9.0', Platform.IOS);
             assert.isFalse(isEnabled);
         });
 
-        it('should return false for Android 7.0 if ab group is 5 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '7.0', Platform.ANDROID);
+        it(`should return false for Android 7.0 if ab group is ${invalidABGroup} and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.`, () => {
+            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(invalidABGroup), '547145938', '7.0', Platform.ANDROID);
             assert.isFalse(isEnabled);
         });
 
-        it('should return false if ab group is 5 and targetGameAppStoreId is not in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '-1', '9.0', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if ab group is 1 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '547145938', '9.0', Platform.ANDROID);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if ab group is 1 and targetGameAppStoreId is not in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '-1', '9.0', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if on Android and os version is 4.x', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '4.8', Platform.ANDROID);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if on iOS and os version is 7.x', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '7.11', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if all conditions do not match', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '-1', '7.9', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-    });
-
-    describe('isSliderEndScreenEnabled', () => {
-        it('should return false for iOS 9.0 if ab group is 5 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '9.0', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false for Android 7.0 if ab group is 5 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '7.0', Platform.ANDROID);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if ab group is 5 and targetGameAppStoreId is not in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '-1', '9.0', Platform.IOS);
+        it(`should return false if ab group is ${invalidABGroup} and targetGameAppStoreId is not in the SliderEndScreenTargetGameIds.`, () => {
+            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(invalidABGroup), '-1', '9.0', Platform.IOS);
             assert.isFalse(isEnabled);
         });
 
@@ -354,172 +312,18 @@ describe('CustomFeatures', () => {
         });
 
         it('should return false if on Android and os version is 4.x', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '4.8', Platform.ANDROID);
+            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(invalidABGroup), '547145938', '4.8', Platform.ANDROID);
             assert.isFalse(isEnabled);
         });
 
         it('should return false if on iOS and os version is 7.x', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '7.11', Platform.IOS);
+            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(invalidABGroup), '547145938', '7.11', Platform.IOS);
             assert.isFalse(isEnabled);
         });
 
         it('should return false if all conditions do not match', () => {
             const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '-1', '7.9', Platform.IOS);
             assert.isFalse(isEnabled);
-        });
-    });
-
-    describe('isSliderEndScreenEnabled', () => {
-        it('should return false for iOS 9.0 if ab group is 5 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '9.0', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false for Android 7.0 if ab group is 5 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '7.0', Platform.ANDROID);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if ab group is 5 and targetGameAppStoreId is not in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '-1', '9.0', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if ab group is 1 and targetGameAppStoreId is in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '547145938', '9.0', Platform.ANDROID);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if ab group is 1 and targetGameAppStoreId is not in the SliderEndScreenTargetGameIds.', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '-1', '9.0', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if on Android and os version is 4.x', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '4.8', Platform.ANDROID);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if on iOS and os version is 7.x', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(5), '547145938', '7.11', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-
-        it('should return false if all conditions do not match', () => {
-            const isEnabled = CustomFeatures.isSliderEndScreenEnabled(toAbGroup(1), '-1', '7.9', Platform.IOS);
-            assert.isFalse(isEnabled);
-        });
-    });
-
-    describe('isParallaxEndScreenEnabled', () => {
-
-        const existingGameId = 500010005;
-        const enabledAbGroups = [toAbGroup(5), toAbGroup(6)];
-        const defaultInnerHeight = 5;
-        const defaultInnerWidth = 3;
-        const defaultUserAgent = 'Mozilla/5.0 (iPhone; CPU iPhone OS 12_3_1 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) GSA/76.0.253539693 Mobile/16F203 Safari/604.1';
-
-        describe('Device model', () => {
-            it('should return false on older Samsung devices', () => {
-                // A5, A3 2015, A3 2016, Note Edge, Young 2, Galaxy Alpha, Galaxy Grand Prime, Galaxy Ace, S5, S4, Pocket Neo
-                const deviceUserAgents = [
-                    'Mozilla/5.0 (Linux; Android 5.0.2; SAMSUNG SM-A500FU Build/LRX22G) AppleWebKit/537.36 (KHTML, like Gecko) SamsungBrowser/3.3 Chrome/38.0.2125.102 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 5.0.2; SAMSUNG SM-A300FU Build/LRX22G) AppleWebKit/537.36 (KHTML, like Gecko) SamsungBrowser/7.2 Chrome/59.0.3071.125 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 5.1.1; SAMSUNG SM-A310F Build/LMY47X) AppleWebKit/537.36 (KHTML, like Gecko) SamsungBrowser/3.5 Chrome/38.0.2125.102 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 6.0.1; SM-N910V Build/MMB29M) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/52.0.2743.98 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 4.4.2; SM-G130H Build/KOT49H) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/30.0.0.0 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 5.0.2; SAMSUNG SM-G850F Build/LRX22G) AppleWebKit/537.36 (KHTML, like Gecko) SamsungBrowser/2.1 Chrome/34.0.1847.76 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; U; Android 4.4.4; en-us; G530 Build/JDQ39) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30',
-                    'Mozilla/5.0 (Linux; Android 4.4.2; SM-G310HN Build/KOT49H) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/30.0.0.0 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; U; Android 4.4.2; tr-tr; SM-G900 Build/KOT49H) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30',
-                    'Mozilla/5.0 (Linux; U; Android 4.1.1; en-us; GT-I19500 Build/JRO03S) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Safari/534.30',
-                    'Mozilla/5.0 (Linux; U; Android 4.1.2; en-gb; GT-S5312 Build/JZO54K) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30'
-                ];
-
-                deviceUserAgents.forEach(deviceUserAgent => {
-                    const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(enabledAbGroups[0], existingGameId, deviceUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                    assert.isFalse(isEnabled);
-                });
-            });
-
-            it('should return true on newer Samsung devices', () => {
-                // S6, S7, S8, S9
-                const deviceUserAgents = [
-                    'Mozilla/5.0 (Linux; Android 6.0.1; SM-G920S Build/MMB29K; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/50.0.2661.86 Mobile Safari/537.36 [FB_IAB/FB4A;FBAV/79.0.0.18.71;]',
-                    'mozilla/5.0 (linux; android 7.1.1; sm-g930 build/kot49h) applewebkit/537.36 (khtml, like gecko) version/4.0 chrome/30.0.0.0 mobile safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 7.0; SM-G950 Build/MMB29K) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/39.0.0.0 Mobile Safari/537.36',
-                    'Mozilla/5.0 (Linux; Android 4.4.2; SM-G960 Build/KOT49H) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/49.0.2623.105 Mobile Safari/537.36'
-                ];
-
-                deviceUserAgents.forEach(deviceUserAgent => {
-                    const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(enabledAbGroups[0], existingGameId, deviceUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                    assert.isTrue(isEnabled);
-                });
-            });
-
-            it('should return true if device is iPhone', () => {
-                const iosUserAgent = 'Mozilla/5.0 (iPhone; CPU iPhone OS 12_3_1 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) GSA/76.0.253539693 Mobile/16F203 Safari/604.1';
-                const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(enabledAbGroups[0], existingGameId, iosUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                assert.isTrue(isEnabled);
-            });
-        });
-
-        describe('Aspect ratio', () => {
-            const tests: {
-                testCase: string;
-                innerHeight: number;
-                innerWidth: number;
-                expectedOutcome: boolean;
-            }[] = [{
-                testCase: 'should return false is aspect ratio is 4:3',
-                innerHeight: 4,
-                innerWidth: 3,
-                expectedOutcome: false
-            }, {
-                testCase: 'should return false if aspect ratio is smaller than 4:3',
-                innerHeight: 3.5,
-                innerWidth: 3,
-                expectedOutcome: false
-            }, {
-                testCase: 'should return true is aspect ratio is bigger than 4:3',
-                innerHeight: 5,
-                innerWidth: 3,
-                expectedOutcome: true
-            }];
-
-            tests.forEach((t) => {
-                it(t.testCase, () => {
-                    assert.equal(CustomFeatures.isParallaxEndScreenEnabled(enabledAbGroups[0], existingGameId, defaultUserAgent, t.innerWidth, t.innerHeight, undefined), t.expectedOutcome);
-                });
-            });
-        });
-
-        describe('Game ID', () => {
-            it('should return false if there are no assets for the given game ID', () => {
-                const nonExistingGameId = -1;
-                const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(enabledAbGroups[0], nonExistingGameId, defaultUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                assert.isFalse(isEnabled);
-            });
-
-            it('should return true if there are assets for the given game ID', () => {
-                const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(enabledAbGroups[0], existingGameId, defaultUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                assert.isTrue(isEnabled);
-            });
-        });
-
-        describe('AB group', () => {
-            it('should return true if experiment is enabled for the given AB group', () => {
-                enabledAbGroups.forEach(abGroup => {
-                    const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(abGroup, existingGameId, defaultUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                    assert.isTrue(isEnabled);
-                });
-            });
-
-            it('should return false if experiment is disabled for the given AB group', () => {
-                const disabledGroup = toAbGroup(16);
-                const isEnabled = CustomFeatures.isParallaxEndScreenEnabled(disabledGroup, existingGameId, defaultUserAgent, defaultInnerWidth, defaultInnerHeight, undefined);
-                assert.isFalse(isEnabled);
-            });
         });
     });
 });