+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Tue Sep 17 19:18:02 UTC 2019
+
+* End Parallax Endcard Experiment [#11182](https://github.com/Applifier/unity-ads-webview/pull/11182)
+* Remove Unused Store Transaction Diagnostics [#11186](https://github.com/Applifier/unity-ads-webview/pull/11186)
+* Add New Fields for CCPA Compliance [#11177](https://github.com/Applifier/unity-ads-webview/pull/11177)
+* Fix AUI Integration Test Stability [#11184](https://github.com/Applifier/unity-ads-webview/pull/11184)
+* Begin Slider v2.1 Experiment on Groups 7 and 8 [#10838](https://github.com/Applifier/unity-ads-webview/pull/10838)
+* Add WWF Load Test on ABGroups 13 and 14 [#11212](https://github.com/Applifier/unity-ads-webview/pull/11212)
+
 # Mon Sep 16 19:23:57 UTC 2019
 
 * Upgrade Akamai Purging Implementation to Supported Version [#11183](https://github.com/Applifier/unity-ads-webview/pull/11183)